requirements-common.txt
sqlalchemy
psycopg2-binary
python-dotenv
pyyaml